+++ modules/services/auto_scaling.tf
@@ -93,12 +93,12 @@ resource "aws_launch_configuration" "launch_config_services" {
 }
 
 data "template_cloudinit_config" "config" {
-  gzip = true
+  gzip          = true
   base64_encode = true
 
   part {
     content_type = "text/cloud-config"
-    content = "${data.template_file.userdata.rendered}"
+    content      = "${data.template_file.userdata.rendered}"
   }
 
   part {