+++ variables.tf
@@ -125,14 +125,14 @@ variable "platform_access_cidrs" {
   description = "A comma delimited list of CIDRs from which to allow access to the site."
 }
 
-variable "ecs_nat_ip" {
+variable "az1_nat_ip" {
   type        = "string"
-  description = "The public IP all traffic from the ECS cluster is NAT'ed through to allow access to the APIs"
+  description = "The public IP all traffic from az1 is NAT'ed through to allow access to the APIs"
 }
 
-variable "services_nat_ip" {
+variable "az2_nat_ip" {
   type        = "string"
-  description = "The public IP all traffic from the Services cluster is NAT'ed through to allow access to the APIs"
+  description = "The public IP all traffic from az2 is NAT'ed through to allow access to the APIs"
 }
 
 variable "ssl_certificate_arn" {